Fix timeout calculations so they are as accurate as the timer code allows (previously they were only accurate to the nearest second).
Patch by me in PR#12418.

[[originally from CVS; cvs2svn created svn r9943]]

+++ common/packets_gen.h
@@ -109,7 +109,7 @@ struct packet_game_info {
   int tech;
   int researchcost;
   int skill_level;
-  int seconds_to_phasedone;
+  float seconds_to_phasedone;
   int timeout;
   int turn;
   int phase;
